+++ OCLEncodeDecode.h
@@ -30,7 +30,7 @@ public:
 	tDeviceRC unmapOutput(void* mappedPtr);
 	void finish(void);
 protected:
-	void run(std::vector<T*> components,int w,int h, int levels);
+	void run(std::vector<T*> components,size_t w,size_t h, size_t levels);
 	ocl_args_d_t* _ocl;
 	bool lossy;
 	OCLMemoryManager<T>* memoryManager;
