+++ ocldwt53.cl
@@ -91,12 +91,12 @@ inline int getCorrectedGlobalIdY() {
 
 
 // read pixel from local buffer
-int4 readPixel( LOCAL short*  restrict  src) {
+int4 readPixel( LOCAL int*  restrict  src) {
 	return (int4)(*src, *(src+CHANNEL_BUFFER_SIZE),  *(src+CHANNEL_BUFFER_SIZE_X2),  *(src+CHANNEL_BUFFER_SIZE_X3)) ;
 }
 
 //write pixel to column
-inline void writePixel(int4 pix, LOCAL short*  restrict  dest) {
+inline void writePixel(int4 pix, LOCAL int*  restrict  dest) {
 	*dest = pix.x;
 	dest += CHANNEL_BUFFER_SIZE;
 	*dest = pix.y;
@@ -107,7 +107,7 @@ inline void writePixel(int4 pix, LOCAL short*  restrict  dest) {
 }
 
 // write row to destination
-void writeRowToOutput(LOCAL short* restrict currentScratch, write_only image2d_t odata,  write_only image2d_t odataLL, int firstX, int outputY, int width, int halfWidth){
+void writeRowToOutput(LOCAL int* restrict currentScratch, write_only image2d_t odata,  write_only image2d_t odataLL, size_t firstX, size_t outputY, size_t width, size_t halfWidth){
 
 	int2 posOut = {firstX>>1, outputY};
 	for (int j = 0; j < WIN_SIZE_X; j+=2) {
@@ -133,14 +133,14 @@ void writeRowToOutput(LOCAL short* restrict currentScratch, write_only image2d_t
 
 
 // initial scratch offset when transforming vertically
-inline int getScratchOffset(){
+inline size_t getScratchOffset(){
    return (getLocalId(1)>> 1) + (getLocalId(1)&1) * BUFFER_SIZE;
 }
 
 // assumptions: width and height are both even
 // (we will probably have to relax these assumptions in the future)
 void KERNEL run(read_only image2d_t idata, write_only image2d_t odataLL, write_only image2d_t odata,
-                       const unsigned int  width, const unsigned int  height, const unsigned int steps) {
+                       const size_t  width, const size_t height, const size_t steps) {
 
 	int inputY = getCorrectedGlobalIdY();
 	int outputY = -1;
@@ -148,7 +148,7 @@ void KERNEL run(read_only image2d_t idata, write_only image2d_t odataLL, write_o
 	    outputY = (inputY >> 1) + (inputY & 1)*( height >> 1);
 
     const unsigned int halfWidth = width >> 1;
-	LOCAL short scratch[PIXEL_BUFFER_SIZE];
+	LOCAL int scratch[PIXEL_BUFFER_SIZE];
 	const float xDelta = 1.0/(width-1);
 	int firstX = getGlobalId(0) * (steps * WIN_SIZE_X);
 
@@ -175,7 +175,7 @@ void KERNEL run(read_only image2d_t idata, write_only image2d_t odataLL, write_o
 	for (int i = 0; i < steps; ++i) {
 
 		// 1. read from source image, transform columns, and store in local scratch
-		LOCAL short* currentScratch = scratch + getScratchOffset();
+		LOCAL int* currentScratch = scratch + getScratchOffset();
 		for (int j = 0; j < WIN_SIZE_X; j+=2) {
 	   
 			///////////////////////////////////////////////////////////////////////////////////////////
